[backend] Refactor to have only one #4639 function. Introduce engine-loader.ts to split the concern and try to move to typescript
