+++ opencti-platform/opencti-graphql/src/domain/retentionRule.js
@@ -2,7 +2,8 @@ import { deleteElementById, updateAttribute } from '../database/middleware';
 import { listEntities, storeLoadById } from '../database/middleware-loader';
 import { ENTITY_TYPE_RETENTION_RULE } from '../schema/internalObject';
 import { generateInternalId, generateStandardId } from '../schema/identifier';
-import { elIndex, elPaginate } from '../database/engine';
+import { elIndex } from '../database/engine';
+import { elPaginate } from '../database/engine-loader';
 import { INDEX_INTERNAL_OBJECTS, READ_DATA_INDICES_WITHOUT_INFERRED } from '../database/utils';
 import { UnsupportedError } from '../config/errors';
 import { utcDate } from '../utils/format';
