+++ opencti-platform/opencti-graphql/src/manager/expiredManager.js
@@ -1,7 +1,7 @@
 import { clearIntervalAsync, setIntervalAsync } from 'set-interval-async/fixed';
 import { Promise } from 'bluebird';
 import { lockResource } from '../database/redis';
-import { elList, ES_MAX_CONCURRENCY } from '../database/engine';
+import { elList, ES_MAX_CONCURRENCY } from '../database/engine-loader';
 import { READ_DATA_INDICES, READ_INDEX_INTERNAL_OBJECTS } from '../database/utils';
 import { prepareDate } from '../utils/format';
 import { patchAttribute } from '../database/middleware';
