+++ opencti-platform/opencti-graphql/src/manager/ruleManager.ts
@@ -27,7 +27,8 @@ import { ALREADY_DELETED_ERROR, TYPE_LOCK_ERROR } from '../config/errors';
 import { getParentTypes } from '../schema/schemaUtils';
 import { isBasicRelationship } from '../schema/stixRelationship';
 import { isStixSightingRelationship } from '../schema/stixSightingRelationship';
-import { elList, internalLoadById } from '../database/middleware-loader';
+import { internalLoadById } from '../database/middleware-loader';
+import { elList } from '../database/engine-loader';
 import type { RuleDefinition, RuleRuntime, RuleScope } from '../types/rules';
 import type { BasicManagerEntity, BasicStoreCommon, BasicStoreEntity, StoreObject } from '../types/store';
 import type { AuthContext, AuthUser } from '../types/user';
@@ -187,7 +188,7 @@ const applyCleanupOnDependencyIds = async (deletionIds: Array<string>) => {
     // eslint-disable-next-line @typescript-eslint/no-use-before-define
     await rulesCleanHandler(context, RULE_MANAGER_USER, elements, RULES_DECLARATION, deletionIds);
   };
-  await elList<BasicStoreCommon>(context, RULE_MANAGER_USER, READ_DATA_INDICES, { filters, callback });
+  await elList(context, RULE_MANAGER_USER, READ_DATA_INDICES, { filters, callback });
 };
 
 export const rulesApplyHandler = async (context: AuthContext, user: AuthUser, events: Array<DataEvent>, forRules: Array<RuleRuntime> = []) => {
